skeleton of async_utils that starts to work with scope
